Minimize unnecessary functions (Remove GPL infringing code)
