+++ src/openvslam/publish/frame_publisher.cc
@@ -133,14 +133,14 @@ unsigned int frame_publisher::draw_tracked_points(cv::Mat& img, const std::vecto
     return num_tracked;
 }
 
-void frame_publisher::update(tracking_module* tracker) {
+void frame_publisher::update(tracking_module* tracker, const cv::Mat& img, double elapsed_ms) {
     std::lock_guard<std::mutex> lock(mtx_);
 
-    tracker->img_gray_.copyTo(img_);
+    img.copyTo(img_);
 
     const auto num_curr_keypts = tracker->curr_frm_.num_keypts_;
     curr_keypts_ = tracker->curr_frm_.keypts_;
-    elapsed_ms_ = tracker->elapsed_ms_;
+    elapsed_ms_ = elapsed_ms;
     mapping_is_enabled_ = tracker->get_mapping_module_status();
     tracking_state_ = tracker->last_tracking_state_;
