+++ src/openvslam/tracking_module.cc
@@ -7,10 +7,8 @@
 #include "openvslam/data/landmark.h"
 #include "openvslam/data/map_database.h"
 #include "openvslam/data/bow_database.h"
-#include "openvslam/feature/orb_extractor.h"
 #include "openvslam/match/projection.h"
 #include "openvslam/module/local_map_updater.h"
-#include "openvslam/util/image_converter.h"
 #include "openvslam/util/yaml.h"
 
 #include <chrono>
@@ -20,39 +18,6 @@
 
 namespace {
 using namespace openvslam;
-
-feature::orb_params get_orb_params(const YAML::Node& yaml_node) {
-    spdlog::debug("load ORB parameters");
-    try {
-        return feature::orb_params(yaml_node);
-    }
-    catch (const std::exception& e) {
-        spdlog::error("failed in loading ORB parameters: {}", e.what());
-        throw;
-    }
-}
-
-double get_true_depth_thr(const camera::base* camera, const YAML::Node& yaml_node) {
-    spdlog::debug("load depth threshold");
-    double true_depth_thr = 40.0;
-    if (camera->setup_type_ == camera::setup_type_t::Stereo || camera->setup_type_ == camera::setup_type_t::RGBD) {
-        const auto depth_thr_factor = yaml_node["depth_threshold"].as<double>(40.0);
-        true_depth_thr = camera->true_baseline_ * depth_thr_factor;
-    }
-    return true_depth_thr;
-}
-
-double get_depthmap_factor(const camera::base* camera, const YAML::Node& yaml_node) {
-    spdlog::debug("load depthmap factor");
-    double depthmap_factor = 1.0;
-    if (camera->setup_type_ == camera::setup_type_t::RGBD) {
-        depthmap_factor = yaml_node["depthmap_factor"].as<double>(depthmap_factor);
-    }
-    if (depthmap_factor < 0.) {
-        throw std::runtime_error("depthmap_factor must be greater than 0");
-    }
-    return depthmap_factor;
-}
 
 double get_reloc_distance_threshold(const YAML::Node& yaml_node) {
     spdlog::debug("load maximum distance threshold where close keyframes could be found");
@@ -76,10 +41,9 @@ double get_use_robust_matcher_for_relocalization_request(const YAML::Node& yaml_
 
 namespace openvslam {
 
-tracking_module::tracking_module(const std::shared_ptr<config>& cfg, system* system, data::map_database* map_db,
+tracking_module::tracking_module(const std::shared_ptr<config>& cfg, double true_depth_thr, system* system, data::map_database* map_db,
                                  data::bow_vocabulary* bow_vocab, data::bow_database* bow_db)
-    : camera_(cfg->camera_), true_depth_thr_(get_true_depth_thr(camera_, util::yaml_optional_ref(cfg->yaml_node_, "Tracking"))),
-      depthmap_factor_(get_depthmap_factor(camera_, util::yaml_optional_ref(cfg->yaml_node_, "Tracking"))),
+    : camera_(cfg->camera_),
       reloc_distance_threshold_(get_reloc_distance_threshold(util::yaml_optional_ref(cfg->yaml_node_, "Tracking"))),
       reloc_angle_threshold_(get_reloc_angle_threshold(util::yaml_optional_ref(cfg->yaml_node_, "Tracking"))),
       enable_auto_relocalization_(get_enable_auto_relocalization(util::yaml_optional_ref(cfg->yaml_node_, "Tracking"))),
@@ -89,30 +53,11 @@ tracking_module::tracking_module(const std::shared_ptr<config>& cfg, system* sys
       frame_tracker_(camera_, 10),
       relocalizer_(bow_db_, util::yaml_optional_ref(cfg->yaml_node_, "Relocalizer")),
       pose_optimizer_(),
-      keyfrm_inserter_(cfg->camera_->setup_type_, true_depth_thr_, map_db, bow_db, 0, cfg->camera_->fps_) {
+      keyfrm_inserter_(cfg->camera_->setup_type_, true_depth_thr, map_db, bow_db, 0, cfg->camera_->fps_) {
     spdlog::debug("CONSTRUCT: tracking_module");
-
-    feature::orb_params orb_params = get_orb_params(util::yaml_optional_ref(cfg->yaml_node_, "Feature"));
-    const auto tracking_params = util::yaml_optional_ref(cfg->yaml_node_, "Tracking");
-    const auto max_num_keypoints = tracking_params["max_num_keypoints"].as<unsigned int>(2000);
-    extractor_left_ = new feature::orb_extractor(max_num_keypoints, orb_params);
-    if (camera_->setup_type_ == camera::setup_type_t::Monocular) {
-        const auto ini_max_num_keypoints = tracking_params["ini_max_num_keypoints"].as<unsigned int>(2 * extractor_left_->get_max_num_keypoints());
-        ini_extractor_left_ = new feature::orb_extractor(ini_max_num_keypoints, orb_params);
-    }
-    if (camera_->setup_type_ == camera::setup_type_t::Stereo) {
-        extractor_right_ = new feature::orb_extractor(max_num_keypoints, orb_params);
-    }
 }
 
 tracking_module::~tracking_module() {
-    delete extractor_left_;
-    extractor_left_ = nullptr;
-    delete extractor_right_;
-    extractor_right_ = nullptr;
-    delete ini_extractor_left_;
-    ini_extractor_left_ = nullptr;
-
     spdlog::debug("DESTRUCT: tracking_module");
 }
 
@@ -141,81 +86,6 @@ std::vector<cv::KeyPoint> tracking_module::get_initial_keypoints() const {
 
 std::vector<int> tracking_module::get_initial_matches() const {
     return initializer_.get_initial_matches();
-}
-
-std::shared_ptr<Mat44_t> tracking_module::track_monocular_image(const cv::Mat& img, const double timestamp, const cv::Mat& mask) {
-    const auto start = std::chrono::system_clock::now();
-
-    // color conversion
-    img_gray_ = img;
-    util::convert_to_grayscale(img_gray_, camera_->color_order_);
-
-    // create current frame object
-    if (tracking_state_ == tracker_state_t::NotInitialized || tracking_state_ == tracker_state_t::Initializing) {
-        curr_frm_ = data::frame(img_gray_, timestamp, ini_extractor_left_, bow_vocab_, camera_, true_depth_thr_, mask);
-    }
-    else {
-        curr_frm_ = data::frame(img_gray_, timestamp, extractor_left_, bow_vocab_, camera_, true_depth_thr_, mask);
-    }
-
-    track();
-
-    const auto end = std::chrono::system_clock::now();
-    elapsed_ms_ = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-
-    std::shared_ptr<Mat44_t> cam_pose_wc = nullptr;
-    if (curr_frm_.cam_pose_cw_is_valid_) {
-        cam_pose_wc = std::allocate_shared<Mat44_t>(Eigen::aligned_allocator<Mat44_t>(), curr_frm_.get_cam_pose_inv());
-    }
-    return cam_pose_wc;
-}
-
-std::shared_ptr<Mat44_t> tracking_module::track_stereo_image(const cv::Mat& left_img_rect, const cv::Mat& right_img_rect, const double timestamp, const cv::Mat& mask) {
-    const auto start = std::chrono::system_clock::now();
-
-    // color conversion
-    img_gray_ = left_img_rect;
-    cv::Mat right_img_gray = right_img_rect;
-    util::convert_to_grayscale(img_gray_, camera_->color_order_);
-    util::convert_to_grayscale(right_img_gray, camera_->color_order_);
-
-    // create current frame object
-    curr_frm_ = data::frame(img_gray_, right_img_gray, timestamp, extractor_left_, extractor_right_, bow_vocab_, camera_, true_depth_thr_, mask);
-
-    track();
-
-    const auto end = std::chrono::system_clock::now();
-    elapsed_ms_ = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-
-    std::shared_ptr<Mat44_t> cam_pose_wc = nullptr;
-    if (curr_frm_.cam_pose_cw_is_valid_) {
-        cam_pose_wc = std::allocate_shared<Mat44_t>(Eigen::aligned_allocator<Mat44_t>(), curr_frm_.get_cam_pose_inv());
-    }
-    return cam_pose_wc;
-}
-
-std::shared_ptr<Mat44_t> tracking_module::track_RGBD_image(const cv::Mat& img, const cv::Mat& depthmap, const double timestamp, const cv::Mat& mask) {
-    const auto start = std::chrono::system_clock::now();
-
-    // color and depth scale conversion
-    img_gray_ = img;
-    cv::Mat img_depth = depthmap;
-    util::convert_to_grayscale(img_gray_, camera_->color_order_);
-    util::convert_to_true_depth(img_depth, depthmap_factor_);
-
-    // create current frame object
-    curr_frm_ = data::frame(img_gray_, img_depth, timestamp, extractor_left_, bow_vocab_, camera_, true_depth_thr_, mask);
-
-    track();
-
-    const auto end = std::chrono::system_clock::now();
-    elapsed_ms_ = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-
-    std::shared_ptr<Mat44_t> cam_pose_wc = nullptr;
-    if (curr_frm_.cam_pose_cw_is_valid_) {
-        cam_pose_wc = std::allocate_shared<Mat44_t>(Eigen::aligned_allocator<Mat44_t>(), curr_frm_.get_cam_pose_inv());
-    }
-    return cam_pose_wc;
 }
 
 bool tracking_module::request_relocalize_by_pose(const Mat44_t& pose) {
@@ -281,11 +151,10 @@ void tracking_module::reset() {
     tracking_state_ = tracker_state_t::NotInitialized;
 }
 
-void tracking_module::track() {
+std::shared_ptr<Mat44_t> tracking_module::track(data::frame curr_frm) {
     if (tracking_state_ == tracker_state_t::NotInitialized) {
         tracking_state_ = tracker_state_t::Initializing;
     }
-
     last_tracking_state_ = tracking_state_;
 
     // check if pause is requested
@@ -294,12 +163,14 @@ void tracking_module::track() {
         std::this_thread::sleep_for(std::chrono::microseconds(5000));
     }
 
+    curr_frm_ = std::move(curr_frm);
+
     // LOCK the map database
     std::lock_guard<std::mutex> lock(data::map_database::mtx_database_);
 
     if (tracking_state_ == tracker_state_t::Initializing) {
         if (!initialize()) {
-            return;
+            return nullptr;
         }
 
         // pass all of the keyframes to the mapping module
@@ -348,7 +219,7 @@ void tracking_module::track() {
             && curr_frm_.id_ - initializer_.get_initial_frame_id() < camera_->fps_ * init_retry_thr) {
             spdlog::info("tracking lost within {} sec after initialization", init_retry_thr);
             system_->request_reset();
-            return;
+            return nullptr;
         }
 
         // show message if tracking has been lost
@@ -377,6 +248,12 @@ void tracking_module::track() {
 
     // update last frame
     last_frm_ = curr_frm_;
+
+    std::shared_ptr<Mat44_t> cam_pose_wc = nullptr;
+    if (curr_frm_.cam_pose_cw_is_valid_) {
+        cam_pose_wc = std::allocate_shared<Mat44_t>(Eigen::aligned_allocator<Mat44_t>(), curr_frm_.get_cam_pose_inv());
+    }
+    return cam_pose_wc;
 }
 
 bool tracking_module::initialize() {
